[REFACTOR] Change file name to camel case to keep consistency
